(PC-31176) refactor(booking): extract SmallBadgedButton in external component
